requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
pyee==9.0.4
python-dotenv==0.20.0
requests==2.27.1
slack-bolt==1.13.2
slack-sdk==3.16.1
slackclient==2.9.4
slackeventsapi==3.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
